id type changed

+++ src/main/java/com/davinci/custockspringboot/domain/dto/stock/CreateBatchDto.java
@@ -7,5 +7,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class CreateBatchDto {
     private String name;
-    private long supplier;
+    private int supplier;
 }
